requirements.txt
aiofiles==0.8.0
aiogram==3.0.0b4
aiohttp==3.8.4
aiohttp-jinja2==1.5.1
aiohttp-security==0.4.0
aiohttp-socks==0.8.0
aiosignal==1.2.0
APScheduler==3.9.1
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.1.0
Babel==2.9.1
beautifulsoup4==4.11.1
boltons==19.3.0
cachetools==4.2.4
certifi==2023.5.7
charset-normalizer==2.1.1
chess==1.10.0
click==8.1.3
colorama==0.4.5
coloredlogs==15.0.1
dataclass-factory==2.16
Deprecated==1.2.14
diagrams==0.23.3
distro==1.8.0
dnspython==2.2.1
environs==9.5.0
frozenlist==1.3.1
fs-cdek-api==1.4.1
fs-pochta-api==1.2.1
geographiclib==2.0
geopy==2.3.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
graphviz==0.20.1
gspread==5.10.0
httplib2==0.22.0
humanfriendly==10.0
idna==3.4
isodate==0.6.1
Jinja2==3.1.2
libretranslatepy==2.1.1
lxml==4.9.1
magic-filter==1.0.9
MarkupSafe==2.1.2
marshmallow==3.17.1
multidict==6.0.2
netaddr==0.8.0
numpy==1.24.3
oauth2client==4.1.3
oauthlib==3.2.2
packaging==21.3
platformdirs==3.6.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.11
Pygments==2.13.0
pymongo==4.2.0
pyparsing==3.0.9
pyreadline3==3.4.1
python-chess==1.999
python-dotenv==0.21.0
python-socks==2.3.0
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rsa==4.9
shapely==2.0.1
six==1.16.0
soupsieve==2.3.2.post1
translate==3.6.1
typed-ast==1.5.4
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
websockets==11.0.3
wrapt==1.15.0
yarl==1.8.1
yookassa==2.4.0
zeep==4.2.1
